+++ plugin/table.go
@@ -29,13 +29,13 @@ type Table struct {
 	// definitions of dependencies between hydrate functions
 	HydrateDependencies []HydrateDependencies
 	HydrateConfig       []HydrateConfig
+	// map of hydrate function name to columns it provides
+	hydrateColumnMap map[string][]string
 }
 
 type GetConfig struct {
 	// key or keys which are used to uniquely identify rows - used to determine whether  a query is a 'get' call
-	KeyColumns         *KeyColumnSet
-	OptionalKeyColumns *KeyColumnSet
-	ItemFromKey        HydrateFunc
+	KeyColumns KeyColumnSlice
 	// the hydrate function which is called first when performing a 'get' call.
 	// if this returns 'not found', no further hydrate functions are called
 	Hydrate HydrateFunc
@@ -45,8 +45,7 @@ type GetConfig struct {
 }
 
 type ListConfig struct {
-	KeyColumns         *KeyColumnSet
-	OptionalKeyColumns *KeyColumnSet
+	KeyColumns KeyColumnSlice
 	// the list function, this should stream the list results back using the QueryData object, and return nil
 	Hydrate HydrateFunc
 	// the parent list function - if we list items with a parent-child relationship, this will list the parent items
@@ -58,38 +57,54 @@ type ListConfig struct {
 // build a list of required hydrate function calls which must be executed, based on the columns which have been requested
 // NOTE: 'get' and 'list' calls are hydration functions, but they must be omitted from this list as they are called
 // first. BEFORE the other hydration functions
+// NOTE2: this function also populates the resolvedHydrateName for each column (used to retrieve column values),
+// and the hydrateColumnMap (used to determine which columns to return)
 func (t *Table) requiredHydrateCalls(colsUsed []string, fetchType fetchType) []*HydrateCall {
 	log.Printf("[TRACE] requiredHydrateCalls, table '%s' fetchType %s colsUsed %v\n", t.Name, fetchType, colsUsed)
 
 	// what is the name of the fetch call (i.e. the get/list call)
-	var fetchCallName string
-	if fetchType == fetchTypeList {
-		fetchCallName = helpers.GetFunctionName(t.List.Hydrate)
-	} else {
-		fetchCallName = helpers.GetFunctionName(t.Get.Hydrate)
-	}
+	fetchFunc := t.getFetchFunc(fetchType)
+	fetchCallName := helpers.GetFunctionName(fetchFunc)
 
+	// initialise hydrateColumnMap
+	t.hydrateColumnMap = make(map[string][]string)
 	requiredCallBuilder := newRequiredHydrateCallBuilder(t, fetchCallName)
 
 	// populate a map keyed by function name to ensure we only store each hydrate function once
 	for _, column := range t.Columns {
-		if helpers.StringSliceContains(colsUsed, column.Name) {
+		// see if this column specifies a hydrate function
 
-			// see if this column specifies a hydrate function
-			hydrateFunc := column.Hydrate
-			if hydrateFunc != nil {
-				hydrateName := helpers.GetFunctionName(hydrateFunc)
-				column.resolvedHydrateName = hydrateName
+		var hydrateName string
+		if hydrateFunc := column.Hydrate; hydrateFunc == nil {
+			// so there is NO hydrate call registered for the column
+			// the column is provided by the fetch call
+			// do not add to map of hydrate functions as the fetch call will always be called
+			hydrateFunc = fetchFunc
+			hydrateName = fetchCallName
+		} else {
+			// there is a hydrate call registered
+			hydrateName = helpers.GetFunctionName(hydrateFunc)
+			// if this column was requested in query, add the hydrate call to required calls
+			if helpers.StringSliceContains(colsUsed, column.Name) {
 				requiredCallBuilder.Add(hydrateFunc)
-			} else {
-				column.resolvedHydrateName = fetchCallName
-				// so there is no hydrate call registered for the column - the resolvedHydrateName is the fetch call
-				// do not add to map of hydrate functions as the fetch call will always be called
 			}
 		}
-	}
 
+		// now update hydrateColumnMap
+		t.hydrateColumnMap[hydrateName] = append(t.hydrateColumnMap[hydrateName], column.Name)
+		// store the hydrate name in the column object
+		column.resolvedHydrateName = hydrateName
+	}
 	return requiredCallBuilder.Get()
+}
+
+func (t *Table) getFetchFunc(fetchType fetchType) HydrateFunc {
+
+	if fetchType == fetchTypeList {
+		return t.List.Hydrate
+	}
+	return t.Get.Hydrate
+
 }
 
 func (t *Table) getHydrateDependencies(hydrateFuncName string) []HydrateFunc {
@@ -118,14 +133,4 @@ func (t *Table) getHydrateConfig(hydrateFuncName string) *HydrateConfig {
 	}
 
 	return config
-}
-
-// return the column matching the given nam
-func (t *Table) columnForName(columnName string) (*Column, bool) {
-	for _, c := range t.Columns {
-		if c.Name == columnName {
-			return c, true
-		}
-	}
-	return nil, false
 }